+++ src/main/kotlin/com/github/gplassard/managedexcludes/dialog/DebugDialog.kt
@@ -1,9 +1,11 @@
 package com.github.gplassard.managedexcludes.dialog
 
 import com.github.gplassard.managedexcludes.settings.PluginSettingsState
+import com.intellij.openapi.project.BaseProjectDirectories.Companion.getBaseDirectories
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.ui.DialogWrapper
 import com.intellij.openapi.vfs.VirtualFile
+import com.intellij.project.isDirectoryBased
 import com.intellij.ui.dsl.builder.panel
 import java.awt.Dimension
 import javax.swing.JComponent
@@ -24,6 +26,29 @@ class DebugDialog(
         val excludedPaths = state.excludedPaths
         val resolvedExcludedPaths = state.resolveExcludedPaths(project)
         return panel {
+            group("Project") {
+                row {
+                    label("name: ${project.name}")
+                }
+                row {
+                    label("isDefault: ${project.isDefault}")
+                }
+                row {
+                    label("isDirectoryBased: ${project.isDirectoryBased}")
+                }
+                row {
+                    label("basePath: ${project.basePath}")
+                }
+                row {
+                    label("baseDirectories: ${project.getBaseDirectories().joinToString(", ") { it.path }}")
+                }
+                row {
+                    label("workspaceFile: ${project.workspaceFile?.path}")
+                }
+                row {
+                    label("projectFilePath: ${project.projectFile?.path}")
+                }
+            }
             group("Tracked Projects:") {
                 for (p in trackedProjects) {
                     row {